requirements.txt
python-dotenv
smolagents[toolkit]
python-docx
click
streamlit